+++ crates/mds-config/src/scan.rs
@@ -1,11 +1,17 @@
 use serde::{Deserialize, Serialize};
 
 use crate::config_type::ConfigType;
+use io_threads::default_io_threads;
+
+pub mod io_threads;
+pub use io_threads::IoThreads;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct Scan {
     pub service_discovery: bool,
     pub tcp_ports: Option<Vec<u16>>,
+    #[serde(default = "default_io_threads")]
+    pub io_threads: IoThreads,
 }
 
 impl Default for Scan {
@@ -13,6 +19,7 @@ impl Default for Scan {
         Self {
             service_discovery: mds_default::SCAN_SERVICE_DISCOVERY.value,
             tcp_ports: Some(mds_default::SCAN_TCP_PORTS.value.to_vec()),
+            io_threads: IoThreads::Dynamic,
         }
     }
 }
@@ -29,6 +36,11 @@ impl Scan {
                 key: "TCP Ports",
                 val: &mut self.tcp_ports,
                 description: mds_default::SCAN_TCP_PORTS.description,
+            },
+            ConfigType::ScanIoThreads {
+                key: "I/O Threads",
+                val: &mut self.io_threads,
+                description: mds_default::SCAN_IO_THREADS.description,
             },
         ]
     }